pyproject.toml
[project]
name = "neodigit-dyndns"
version = "0.3.3"
description = "A dynamic DNS client using Neodigit's API"
authors = [{name = "Eduardo Collado", email = "edu@tecnocratica.net"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.24.0",
    "PyYAML>=5.3.1",
]

[project.scripts]
neodigit-dyndns = "neodigit_dyndns.__main__:main"

[project.urls]
Homepage = "https://github.com/educollado/neodigit-dyndns"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
